mm-video-legacy: venc: update metadata-type used for HAL1 based  recording
Update metadata buffer type from kMetadataBufferTypeCameraSource to
kMetadataBufferTypeNativeHandleSource as the latter is used by
camera-HALv1 that resides in a separate process.

Surface recording uses a different metadata as compared to
legacy camera metadata (HAL1).Cast the gralloc-meta to
appropriate struct to extract the handle

Since buffer-size is not known conclusively at the allocation
time of meta-handles, allow the least size to be allocated.
But check the buffer size before accessing in ETB

Bug: 26268807

Adapted from:
https://github.com/RonGokhale/android_hardware_qcom_media/commit/7bd1c12e14aa249d30ad18cff33f00282a485141

Change-Id: Idfa4211f468aa622170ca7a789cfbb25f31e594d